+++ randovania/layout/preset.py
@@ -1,5 +1,7 @@
+from __future__ import annotations
+
 import dataclasses
-import uuid
+import uuid as uuid_module
 from typing import Iterator
 
 from randovania.bitpacking.bitpacking import BitPackDecoder, BitPackValue
@@ -10,9 +12,8 @@ from randovania.layout.base.base_configuration import BaseConfiguration
 @dataclasses.dataclass(frozen=True)
 class Preset(BitPackValue):
     name: str
-    uuid: uuid.UUID
+    uuid: UUID
     description: str
-    base_preset_uuid: uuid.UUID | None
     game: RandovaniaGame
     configuration: BaseConfiguration
 
@@ -22,19 +23,17 @@ class Preset(BitPackValue):
             "name": self.name,
             "uuid": str(self.uuid),
             "description": self.description,
-            "base_preset_uuid": str(self.base_preset_uuid) if self.base_preset_uuid is not None else None,
             "game": self.game.value,
             "configuration": self.configuration.as_json,
         }
 
     @classmethod
-    def from_json_dict(cls, value) -> "Preset":
+    def from_json_dict(cls, value) -> Preset:
         game = RandovaniaGame(value["game"])
         return Preset(
             name=value["name"],
-            uuid=uuid.UUID(value["uuid"]),
+            uuid=uuid_module.UUID(value["uuid"]),
             description=value["description"],
-            base_preset_uuid=uuid.UUID(value["base_preset_uuid"]) if value["base_preset_uuid"] is not None else None,
             game=game,
             configuration=game.data.layout.configuration.from_json(value["configuration"]),
         )
@@ -45,7 +44,7 @@ class Preset(BitPackValue):
     def settings_incompatible_with_multiworld(self) -> list[str]:
         return self.configuration.settings_incompatible_with_multiworld()
 
-    def is_same_configuration(self, other: "Preset") -> bool:
+    def is_same_configuration(self, other: Preset) -> bool:
         return self.configuration == other.configuration
 
     def bit_pack_encode(self, metadata) -> Iterator[tuple[int, int]]:
@@ -56,7 +55,7 @@ class Preset(BitPackValue):
         yield from self.configuration.bit_pack_encode({"reference": reference.configuration})
 
     @classmethod
-    def bit_pack_unpack(cls, decoder: BitPackDecoder, metadata) -> "Preset":
+    def bit_pack_unpack(cls, decoder: BitPackDecoder, metadata) -> Preset:
         from randovania.interface_common.preset_manager import PresetManager
         manager: PresetManager = metadata["manager"]
         game: RandovaniaGame = metadata["game"]
@@ -66,13 +65,14 @@ class Preset(BitPackValue):
         return Preset(
             name=f"{game.long_name} Custom",
             description="A customized preset.",
-            uuid=uuid.uuid4(),
-            base_preset_uuid=reference.uuid,
+            uuid=uuid_module.uuid4(),
             game=reference.game,
             configuration=reference.configuration.bit_pack_unpack(decoder, {"reference": reference.configuration}),
         )
 
     def fork(self) -> "Preset":
-        return dataclasses.replace(self, name=f"{self.name} Copy",
-                                   description=f"A copy version of {self.name}.",
-                                   uuid=uuid.uuid4(), base_preset_uuid=self.uuid)
+        return dataclasses.replace(
+            self, name=f"{self.name} Copy",
+            description=f"A copy version of {self.name}.",
+            uuid=uuid_module.uuid4(),
+        )