+++ test/games/prime1/gui/preset_settings/test_prime_enemy_stat_randomizer.py
@@ -15,9 +15,7 @@ from randovania.games.game import RandovaniaGame
 def test_on_preset_changed(skip_qtbot, preset_manager, use_enemy_attribute_randomizer):
     #Setup
     base = preset_manager.default_preset_for_game(RandovaniaGame.METROID_PRIME).get_preset()
-    preset = dataclasses.replace(base,
-                                 uuid=uuid.UUID('b41fde84-1f57-4b79-8cd6-3e5a78077fa6'),
-                                 base_preset_uuid=base.uuid)
+    preset = dataclasses.replace(base, uuid=uuid.UUID('b41fde84-1f57-4b79-8cd6-3e5a78077fa6'))
     editor = PresetEditor(preset)
     window = PresetEnemyAttributeRandomizer(editor, default_database.game_description_for(preset.game), MagicMock())